move tests to typescript
