+++ package.json
@@ -8,8 +8,7 @@
     "build": "run-s clean build-ts",
     "build-ts": "tsc",
     "lint": "eslint \"@(src|test)/**/*.ts\"",
-    "pretest": "npm run build",
-    "test": "mocha",
+    "test": "ts-mocha -p tsconfig.test.json test/**/*.ts",
     "test:watch": "mocha --w",
     "prepublish": "npm run build",
     "clean": "rimraf ./lib"
@@ -38,6 +37,7 @@
     "@types/chai": "^4.2.12",
     "@types/chai-dom": "^0.0.10",
     "@types/mocha": "^8.0.0",
+    "@types/node": "^14.0.27",
     "@typescript-eslint/eslint-plugin": "^3.7.1",
     "@typescript-eslint/parser": "^3.7.1",
     "ajv": "^6.12.3",
@@ -52,6 +52,8 @@
     "mocha": "^8.0.1",
     "npm-run-all": "^4.1.5",
     "rimraf": "^3.0.2",
+    "ts-mocha": "^7.0.0",
+    "ts-node": "^8.10.2",
     "typescript": "^3.9.7"
   },
   "dependencies": {
