+++ test/array.ts
@@ -1,18 +1,18 @@
-import chai from 'chai';
+import chai from 'chai'
 
-import empty from '../src/empty';
+import {empty} from '../src/empty'
 
-chai.should();
+chai.should()
 
 describe('array schema definition', () => {
 
   it('should yield empty array', (done) => {
     empty({
       type: 'array'
-    }).should.deep.equal([]);
+    }).should.deep.equal([])
 
-    done();
-  });
+    done()
+  })
 
   it('should work with tuples', (done) => {
     empty({
@@ -21,18 +21,18 @@ describe('array schema definition', () => {
         {type: 'integer'},
         {type: 'string'}
       ]
-    }).should.deep.equal([0, '']);
+    }).should.deep.equal([0, ''])
 
-    done();
-  });
+    done()
+  })
 
   it('should work with minItems', (done) => {
     empty({
       type: 'array',
       items:{type: 'integer'},
       minItems: 5
-    }).should.deep.equal([0, 0, 0, 0, 0]);
+    }).should.deep.equal([0, 0, 0, 0, 0])
 
-    done();
-  });
-});
+    done()
+  })
+})
