+++ tsconfig.test.json
@@ -1,7 +1,7 @@
 {
     "compilerOptions": {
-      "target": "es6",
-      "module": "esnext",
+      "target": "es5",
+      "module": "commonjs",
       "moduleResolution": "node",
       "lib": ["dom", "es2018"],
       "declaration": true,
